requirements.txt
pytest==6.2.2
pytorch-lightning==1.2.6
--find-links https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.0+cu111
torch==1.8.0+cu111
pandas
pydot
torchviz
pyyaml